yearbook CSS changes and superuser restrictions on profile page

+++ myapp/views.py
@@ -101,6 +101,8 @@ def index(request):
 def profile(request):
     email = request.user.email
     exceptions = ['akul.gupta.phe17@itbhu.ac.in','subodh.verma.min19@itbhu.ac.in','vijay.psingh.chy14@itbhu.ac.in']
+    if request.user.is_superuser:
+        return HttpResponse('super users not allowed')
     if not(re.findall("15@", email) or re.findall("16@", email)) and not(email in exceptions):
         user = User.objects.get(email=email)
         user.delete()
